LayoutTests/canvas/philip/tests/2d.imageData.get.basic-expected.txt

< [index] >
2d.imageData.get.basic
getImageData() exists and returns something
References: 2d.imageData.get.basic
Actual output:
Passed
